parser CHANGE internally always work with full decimal64 precision
Added a function for converting decimal64 leaves/leaf-lists
to double.

+++ src/resolve.h
@@ -96,8 +96,8 @@ struct len_ran_intv {
         } sval;
 
         struct {
-            long double min;
-            long double max;
+            int64_t min;
+            int64_t max;
         } fval;
     } value;
